Restructure repository
