+++ Makefile
@@ -1,8 +1,8 @@
 py_dir              := ./src
-py_main             := $(py_dir)/edge-deep-learning.py
+py_jetson_demo      := $(py_dir)/jetson_demo.py
 py_segnet           := $(py_dir)/segnet.py
-py_training         := $(py_dir)/training2.py
-py_training_requirements  := requirements-training.txt
+py_client           := $(py_dir)/client.py
+py_client_requirements  := requirements-client.txt
 py_server           := $(py_dir)/server.py
 py_stats            := $(py_dir)/collect-statistics.py
 py_cache            := $(shell find . -iname __pycache__)
@@ -33,9 +33,9 @@ $(stats_dir):
 
 $(py_venv): $(py_venv)/touchfile
 
-$(py_venv)/touchfile: $(py_training_requirements)
+$(py_venv)/touchfile: $(py_client_requirements)
 	python3 -m venv $(py_venv)
-	$(py_venv)/bin/pip install -r $(py_training_requirements)
+	$(py_venv)/bin/pip install -r $(py_client_requirements)
 	touch $(py_venv)/touchfile
 
 .PHONY: jetson-dependencies
@@ -43,9 +43,9 @@ jetson-dependencies:
 	$(jetson_install_pytorch_script)
 	$(jetson_install_torchvision_script)
 
-.PHONY: run-classification
-run-classification: $(samples_dir)/$(sample_classification)
-	python3 $(py_main) $(samples_dir)/$(sample_classification)
+.PHONY: run-jetson-demo
+run-jetson-demo: $(samples_dir)/$(sample_classification)
+	python3 $(py_jetson_demo) $(samples_dir)/$(sample_classification)
 
 .PHONY: run-segnet
 run-segnet: $(samples_dir)/$(sample_segmentation)
@@ -55,21 +55,21 @@ run-segnet: $(samples_dir)/$(sample_segmentation)
 run-server:
 	python3 $(py_server)
 
-.PHONY: run-training
-run-training:
-	python3 $(py_training)
+.PHONY: run-client
+run-client:
+	python3 $(py_client)
 
-.PHONY: run-training-venv
-run-training-venv: $(py_venv) $(py_training_requirements)
-	$(py_venv)/bin/python $(py_training)
+.PHONY: run-client-venv
+run-client-venv: $(py_venv) $(py_client_requirements)
+	$(py_venv)/bin/python $(py_client)
 
 .PHONY: run
 run:
-	make run-training
+	make run-client
 
 .PHONY: run-venv
 run-venv:
-	make run-training-venv
+	make run-client-venv
 
 .PHONY: collect-stats
 collect-stats: $(stats_dir)
