+++ src/models/neural_network.py
@@ -1,4 +1,5 @@
 from torch import nn
+from torch.autograd import Variable
 
 class NeuralNetwork(nn.Module):
     def __init__(self):
@@ -7,12 +8,43 @@ class NeuralNetwork(nn.Module):
         self.linear_relu_stack = nn.Sequential(
             nn.Linear(28*28, 512),
             nn.ReLU(),
-            nn.Linear(512, 512),
+            nn.Linear(512, 256),
             nn.ReLU(),
-            nn.Linear(512, 10)
+            nn.Linear(256, 10)
         )
 
     def forward(self, x):
         x = self.flatten(x)
         logits = self.linear_relu_stack(x)
         return logits
+
+class NeuralNetwork_local(nn.Module):
+    def __init__(self):
+        super(NeuralNetwork_local, self).__init__()
+        self.flatten = nn.Flatten()
+        self.linear_relu_stack = nn.Sequential(
+            nn.Linear(28*28, 512),
+            nn.ReLU()
+        )
+
+    def forward(self, x):
+        x = self.flatten(x)
+        logits = self.linear_relu_stack(x)
+        return logits
+
+
+class NeuralNetwork_server(nn.Module):
+    def __init__(self):
+        super(NeuralNetwork_server, self).__init__()
+        self.flatten = nn.Flatten()
+        self.linear_relu_stack = nn.Sequential(
+            nn.Linear(512, 256),
+            nn.ReLU(),
+            nn.Linear(256, 10)
+        )
+
+    def forward(self, x):
+        x = self.flatten(x)
+        logits = self.linear_relu_stack(x)
+        return logits
+
