+++ src/training.py
@@ -1,81 +1,55 @@
 import torch
 from torch import nn
 
-from models.neural_network import NeuralNetwork
-from datasets.mnist_fashion import load_mnist_fashion_dataset
+from networking import run_offload_training
 
-def get_device():
-    device = "cuda" if torch.cuda.is_available() else "cpu"
-    print(f"\nUsing {device} device")
-    return device
-
-def train(dataloader, model, loss_fn, optimizer):
+def train_epoch(dataloader, model_local, loss_fn, optimizer_local, device = 'cpu'):
     size = len(dataloader.dataset)
-    model.train()
+    model_local.train()
+
     for batch, (X, y) in enumerate(dataloader):
-        X, y = X.to(device), y.to(device)
+        X, y = X.to(device), y.to('cpu')
 
-        # Compute prediction error
-        pred = model(X)
-        loss = loss_fn(pred, y)
+        # Local forward propagation
+        split_vals = model_local(X)
 
-        # Backpropagation
-        optimizer.zero_grad()
-        loss.backward()
-        optimizer.step()
+        # Offloaded layers
+        np_split_grad, loss = run_offload_training(split_vals.detach().numpy(), y.detach().numpy())
+
+        # Local back propagation
+        split_grad = torch.from_numpy(np_split_grad).to('cpu')
+        optimizer_local.zero_grad()
+        split_vals.backward(split_grad)
+        optimizer_local.step()
 
         if batch % 100 == 0:
-            loss, current = loss.item(), batch * len(X)
+            current = batch * len(X)
             print(f"loss: {loss:>7f}  [{current:>5d}/{size:>5d}]")
 
-def test(dataloader, model, loss_fn):
+# TODO: Implement evaluation over network
+def test(dataloader, model1, model2, loss_fn, device = 'cpu'):
     size = len(dataloader.dataset)
     num_batches = len(dataloader)
-    model.eval()
+    model1.eval()
+    model2.eval()
     test_loss, correct = 0, 0
     with torch.no_grad():
         for X, y in dataloader:
             X, y = X.to(device), y.to(device)
-            pred = model(X)
+            pred = model2(model1(X))
             test_loss += loss_fn(pred, y).item()
             correct += (pred.argmax(1) == y).type(torch.float).sum().item()
     test_loss /= num_batches
     correct /= size
-    print(f"Test Error: \n Accuracy: {(100*correct):>0.1f}%, Avg loss: {test_loss:>8f}")
+    print(f"Test Error: \n Accuracy: {(100*correct):>0.1f}%, Avg loss: {test_loss:>8f} \n")
 
-def train_model(model, train_dataloader, loss_fn, optimizer, test_dataloader, epochs = 5):
-    print("\nTraining model")
-    for t in range(epochs):
-        print(f"\nEpoch {t+1}\n-------------------------------")
-        train(train_dataloader, model, loss_fn, optimizer)
-        test(test_dataloader, model, loss_fn)
-    print("Done!")
-
-def evaluate_model(model, dataloader, classes):
-    print("\nEvaluating model")
-    model.eval()
-    x, y = next(iter(dataloader))
-    x, y = x.to(device), y.to(device)
+# TODO: Implement evaluation over network
+def evaluate_model(model1, model2, dataloader):
+    model1.eval()
+    model2.eval()
+    x, y = next(iter(test_dataloader))
     with torch.no_grad():
-        pred = model(x[0])
+        pred = model2(model1(x[0]))
         predicted, actual = classes[pred[0].argmax(0)], classes[y[0]]
-        print(f'Predicted: "{predicted}", Actual: "{actual}"\n')
-
-def save_model(model, filepath):
-    torch.save(model.state_dict(), filepath)
-    print("Saved PyTorch Model State to {:s}".format(filepath))
-
-if __name__ == "__main__":
-    device = get_device()
-
-    model = NeuralNetwork().to(device)
-    train_dataloader, test_dataloader, classes = load_mnist_fashion_dataset()
-
-    train_model(model,
-                train_dataloader,
-                nn.CrossEntropyLoss(),
-                torch.optim.SGD(model.parameters(), lr=1e-3),
-                test_dataloader)
-
-    evaluate_model(model, test_dataloader, classes)
+        print(f'Predicted: "{predicted}", Actual: "{actual}"')
